refactor: move utils to shared folder
