+++ src/shared/equalities/index.ts
