+++ src/shared/equalities/shallowEquals.ts
