+++ src/shared/hocs/deepMemo.ts
@@ -1,6 +1,6 @@
 import { ComponentType } from "react";
 
-import { deepEquals } from "../equalities";
+import { deepEquals } from "../equalities/index.ts";
 import { memo } from "./memo.ts";
 
 export function deepMemo<P extends object>(Component: ComponentType<P>) {
