+++ src/shared/index.ts
@@ -1,3 +1,4 @@
 export * from "./equalities";
 export * from "./hooks";
 export * from "./hocs";
+export * from "./tests";
