+++ src/shared/tests/index.ts
@@ -1,0 +1,2 @@
+export * from "./renderLog";
+export * from "./generateItems";
