+++ src/shared/tests/renderLog.ts
@@ -1,0 +1,3 @@
+export function renderLog(message: string) {
+  console.log(message);
+}
