+++ src/shared/utils/isRecord.ts
