+++ src/widgets/complex-form/ui/index.tsx
@@ -2,7 +2,7 @@ import { memo } from "react";
 
 import { useNotification } from "@features/notification/model";
 
-import { renderLog } from "@/utils";
+import { renderLog } from "@shared/tests/renderLog";
 
 import { useComplexForm } from "../model";
