+++ src/widgets/header/ui/index.tsx
@@ -2,7 +2,7 @@ import { useTheme } from "@app/model";
 import { AuthActions } from "@widgets/auth-actions";
 import { useAuth } from "@features/auth/model/AuthProvider";
 
-import { renderLog } from "@/utils";
+import { renderLog } from "@shared/tests/renderLog";
 
 export function Header() {
   renderLog("Header rendered");
