+++ src/widgets/notification-system/ui/NotificationSystem.tsx
@@ -1,6 +1,6 @@
 import { useNotification } from "@features/notification/model";
 
-import { renderLog } from "@/utils";
+import { renderLog } from "@shared/tests/renderLog";
 
 import { NotificationItem } from "./Notificationitem";
