+++ src/widgets/product-list/ui/ProductList.tsx
@@ -1,7 +1,7 @@
 import { useTheme } from "@app/model";
 import { useProducts } from "@features/product/model/ProductProvider";
 
-import { renderLog } from "@/utils";
+import { renderLog } from "@shared/tests/renderLog";
 
 import { useProductList } from "../model";
 import { ProductItem } from "./ProductItem";
