+++ vite.config.ts
@@ -25,7 +25,7 @@ export default mergeConfig(
     test: {
       globals: true,
       environment: "jsdom",
-      setupFiles: "./src/setupTests.ts",
+      setupFiles: "./src/__tests__/setupTests.ts",
       coverage: {
         reportsDirectory: "./.coverage",
         reporter: ["lcov", "json", "json-summary"],
